# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ebdaa Skill Matrix System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees with optional filters
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: string
        - name: current_class
          in: query
          schema:
            type: string
            enum: [A, B, C]
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Deactivate employee
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}/profile:
    get:
      operationId: getEmployeeProfile
      tags: [employees]
      summary: Get full employee profile with evaluations and training
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeProfile"
  /employees/me/profile:
    get:
      operationId: getMyProfile
      tags: [employees]
      summary: Get the authenticated employee's own profile
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Authenticated employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeProfile"
        "403":
          description: Not an employee account or no linked employee record

  # ── SKILLS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        manager_id:
          type: string
          nullable: true
        manager_email:
          type: string
          nullable: true
        employee_count:
          type: integer
        created_at:
          type: string
      required: [id, name, employee_count, created_at]
...
    Employee:
      type: object
      properties:
        id:
          type: string
        employee_code:
          type: string
          nullable: true
        full_name:
          type: string
        department_id:
          type: string
        job_title:
          type: string
          nullable: true
        joined_date:
          type: string
          nullable: true
        birth_date:
          type: string
          nullable: true
        current_class:
          type: string
          enum: [A, B, C]
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        department:
          $ref: "#/components/schemas/Department"
      required: [id, full_name, department_id, current_class, is_active, created_at, department]
    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [data, total, page, page_size]
    CreateEmployeeBody:
      type: object
      properties:
        employee_code:
          type: string
        full_name:
          type: string
        department_id:
          type: string
        job_title:
          type: string
        joined_date:
          type: string
        birth_date:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
      required: [full_name, department_id]
    EmployeeProfile:
      type: object
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        latest_summary:
          $ref: "#/components/schemas/EvaluationSummary"
          nullable: true
        historical_summaries:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationSummary"
        skill_scores:
          type: array
          items:
            $ref: "#/components/schemas/SkillScore"
        training_recommendations:
          type: array
          items:
            $ref: "#/components/schemas/TrainingRecommendation"
      required: [employee, historical_summaries, skill_scores, training_recommendations]
    SkillScore:
      type: object
      properties:
        skill_id:
          type: string
        skill_name:
          type: string
        skill_code:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        weight:
          type: integer
        criticality:
          type: string
          enum: [Low, Medium, High, Critical]
        score:
          type: integer
          nullable: true
        score_label:
          type: string
          nullable: true
      required: [skill_id, skill_name, weight, criticality]

    # ── SKILLS ────────────────────────────────────────────────────────────────
...
    EvaluationSummary:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
        employee_id:
          type: string
        total_score:
          type: number
        max_possible_score:
          type: number
        percentage:
          type: number
        class:
          type: string
          enum: [A, B, C]
        evaluated_skills_count:
          type: integer
        updated_at:
          type: string
        campaign_title:
          type: string
          nullable: true
        employee_name:
          type: string
          nullable: true
        employee_code:
          type: string
          nullable: true
      required: [id, campaign_id, employee_id, total_score, max_possible_score, percentage, class, evaluated_skills_count, updated_at]
...
    TrainingRecommendation:
      type: object
      properties:
        id:
          type: string
        employee_id:
          type: string
        skill_id:
          type: string
          nullable: true
        campaign_id:
          type: string
          nullable: true
        score:
          type: integer
          nullable: true
        recommendation_type:
          type: string
          enum: [Immediate, Short-term, Long-term, Promotion]
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        assigned_trainer_id:
          type: string
          nullable: true
        target_date:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        employee_name:
          type: string
          nullable: true
        skill_name:
          type: string
          nullable: true
      required: [id, employee_id, recommendation_type, status, created_at]
...